+++ src/index.js
@@ -3,7 +3,7 @@ import { Environment } from './core/environment.js'
 import { TemplatePlugin } from './core/templates.js'
 
 /** Create an environment. See Environment.create. */
-export default function wintersmith(...args) {
+export default function coldsmith(...args) {
   return Environment.create(...args)
 }